# Core/contracts/openapi/daemon-api.yaml
openapi: "3.1.0"
info:
  title: ScientificState Local Daemon API
  version: "0.1.0"
  description: |
    Local execution daemon API — localhost only.
    The daemon performs scientific computation; it does NOT assert scientific claims.
    All compute outputs require researcher review before claim evaluation.

    Constitutional boundaries:
    - P7: Non-delegation of scientific authority — daemon executes, human evaluates
    - P1: Raw data immutability — daemon never mutates stored data
    - Localhost boundary: No remote daemon access in M1
  contact:
    name: ScientificState
    url: https://scientificstate.org
  license:
    name: Constitutional Science Infrastructure License
servers:
  - url: "http://127.0.0.1:9473"
    description: Local daemon (Execution_Plan_Phase0.md §4.1 — port 9473)
...
paths:
...
  /workspaces:
    get:
      operationId: listWorkspaces
      tags: [workspaces]
      summary: List workspaces
      responses:
        "200":
          description: Workspace list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspacesResponse"

    post:
      operationId: createWorkspace
      tags: [workspaces]
      summary: Create a new workspace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWorkspaceRequest"
      responses:
        "201":
          description: Workspace created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceDetail"
  /workspaces/{workspace_id}:
    get:
      operationId: getWorkspace
      tags: [workspaces]
      summary: Get workspace details
      parameters:
        - name: workspace_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Workspace details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceDetail"
        "404":
          description: Workspace not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /runs:
    post:
      operationId: createRun
      tags: [compute]
      summary: Submit a compute run
      description: |
        Submits a domain compute run for execution. Returns immediately with run_id.
        Constitutional constraint: daemon executes — researcher evaluates (P7).
        Compute output is not a claim — it requires human review before claim creation.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ComputeRunRequest"
      responses:
        "202":
          description: Run accepted — execution is asynchronous
          content:
            application/json:
              schema:
                type: object
                required: [run_id]
                properties:
                  run_id:
                    type: string
                    description: Permanent run identifier — use with GET /runs/{run_id}
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Domain or method not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /runs/{run_id}:
    get:
      operationId: getRun
      tags: [compute]
      summary: Get compute run status and result
      parameters:
        - name: run_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Run status and result (if completed)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComputeRunResult"
        "404":
          description: Run not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    WorkspacesResponse:
      type: object
      required: [workspaces]
      properties:
        workspaces:
          type: array
          items:
            $ref: "#/components/schemas/WorkspaceSummary"
    WorkspaceSummary:
      type: object
      required: [workspace_id, name, created_at]
      properties:
        workspace_id:
          type: string
        name:
          type: string
        created_at:
          type: string
          format: date-time
    CreateWorkspaceRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
    WorkspaceDetail:
      type: object
      required: [workspace_id, name, created_at, ssv_count, claim_count]
      properties:
        workspace_id:
          type: string
        name:
          type: string
        description:
          type: string
        created_at:
          type: string
          format: date-time
        ssv_count:
          type: integer
        claim_count:
          type: integer
    ComputeRunRequest:
      type: object
      required: [workspace_id, domain_id, method_id]
      properties:
        workspace_id:
          type: string
          description: Target workspace identifier
        domain_id:
          type: string
          description: Domain module identifier (e.g. "polymer_science")
        method_id:
          type: string
          description: Method identifier within the domain (e.g. "kmd_analysis")
        dataset_ref:
          type: string
          description: raw_data_id from /datasets/ingest response
        assumptions:
          type: array
          items:
            type: object
          description: "P3: explicit scientific assumptions for this run"
        parameters:
          type: object
          description: Method-specific parameters
        compute_class:
          type: string
          enum: [classical, quantum_sim, quantum_hw, hybrid]
          description: "Optional compute class hint — defaults to classical if omitted"
      additionalProperties: false
    ComputeRunResult:
      type: object
      required: [run_id, workspace_id, domain_id, method_id, status, started_at]
      properties:
        run_id:
          type: string
        workspace_id:
          type: string
        domain_id:
          type: string
        method_id:
          type: string
        status:
          type: string
          enum: [pending, running, succeeded, failed]
        started_at:
          type: string
          format: date-time
        finished_at:
          type: string
          format: date-time
        ssv_ref:
          type: string
          description: SSV id — only when status=succeeded
        result:
          type: object
          description: Domain computation output — only when status=succeeded
        execution_witness:
          type: object
          required: [compute_class, backend_id]
          properties:
            compute_class:
              type: string
              enum: [classical, quantum_sim, quantum_hw, hybrid]
            backend_id:
              type: string
        error:
          type: object
          properties:
            error_code:
              type: string
            message:
              type: string
          description: Error detail — only when status=failed
      additionalProperties: false
...
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
        message:
          type: string
        detail:
          type: object